follow Pure.css's guidelines
put the padding in a child div nested inside the parent div with the
pure-u-* class, not on the parent div.

I used view units ("vw", percent of viewport width, which is the window
width) instead of percent for the new padding, otherwise the 5% of a
pure-u-1-2 column did not match the 5% of a full-width pure-u-1.

also, I've discovered that putting two pure-u-1-2 inside a pure-u-1 is
messing things up, better to never nest pure-u classes and to instead
repeat the pure-g as needed.

a side effect of this change is that there is now more space between
columns, because the padding is applied to each child column, not to the
parent div. Still looks good, I think.